Fixed bug with some reduction functors returning floats instead of doubles.

+++ lib/reduce_quda.cu
@@ -131,8 +131,8 @@ namespace quda {
      Return the L1 norm of x
   */
   __device__ double norm1_(const double2 &a) { return fabs(a.x) + fabs(a.y); }
-  __device__ float norm1_(const float2 &a) { return (double)fabs(a.x) + (double)fabs(a.y); }
-  __device__ float norm1_(const float4 &a) { return (double)fabs(a.x) + (double)fabs(a.y) + (double)fabs(a.z) + (double)fabs(a.w); }
+  __device__ double norm1_(const float2 &a) { return (double)fabs(a.x) + (double)fabs(a.y); }
+  __device__ double norm1_(const float4 &a) { return (double)fabs(a.x) + (double)fabs(a.y) + (double)fabs(a.z) + (double)fabs(a.w); }
 
   template <typename ReduceType, typename Float2, typename FloatN>
 #if (__COMPUTE_CAPABILITY__ >= 200)
@@ -161,8 +161,8 @@ namespace quda {
      Return the L2 norm of x
   */
   __device__ double norm2_(const double2 &a) { return a.x*a.x + a.y*a.y; }
-  __device__ float norm2_(const float2 &a) { return (double)a.x*(double)a.x + (double)a.y*(double)a.y; }
-  __device__ float norm2_(const float4 &a) { return (double)a.x*(double)a.x + (double)a.y*(double)a.y +
+  __device__ double norm2_(const float2 &a) { return (double)a.x*(double)a.x + (double)a.y*(double)a.y; }
+  __device__ double norm2_(const float4 &a) { return (double)a.x*(double)a.x + (double)a.y*(double)a.y +
       (double)a.z*(double)a.z + (double)a.w*(double)a.w; }
 
   template <typename ReduceType, typename Float2, typename FloatN>
@@ -187,8 +187,8 @@ namespace quda {
      Return the real dot product of x and y
   */
     __device__ double dot_(const double2 &a, const double2 &b) { return a.x*b.x + a.y*b.y; }
-    __device__ float dot_(const float2 &a, const float2 &b) { return (double)a.x*(double)b.x + (double)a.y*(double)b.y; }
-    __device__ float dot_(const float4 &a, const float4 &b) { return (double)a.x*(double)b.x + (double)a.y*(double)b.y + (double)a.z*(double)b.z + (double)a.w*(double)b.w; }
+    __device__ double dot_(const float2 &a, const float2 &b) { return (double)a.x*(double)b.x + (double)a.y*(double)b.y; }
+    __device__ double dot_(const float4 &a, const float4 &b) { return (double)a.x*(double)b.x + (double)a.y*(double)b.y + (double)a.z*(double)b.z + (double)a.w*(double)b.w; }
 
   template <typename ReduceType, typename Float2, typename FloatN>
 #if (__COMPUTE_CAPABILITY__ >= 200)
